Update patch set 1
Patch Set 1:

This is adding "iw" to the AOSP manifest.

We already fixed Android.mk in that project to build correctly.

I followed the instructions in b/22537539#comment21 I hope I got everything correct. In particular, not sure if I needed a groups="..." attribute for this <project/>, let me know if that's the case and I'll update the commit.

Cheers,
Filipe

Patch-set: 1