Update patch set 1
Patch Set 1:

> brillo-specific projects should in addition go into brillo/manifest
 > as well. Adding it here is fine because we want to make sure the
 > project gets checkbuilded and merged internally.

Some background on why I added this to the Android manifest only for now: The current dev environment for Trusty-based access-controlled NVRAM is N9 / flounder, so having it in the Android manifest.xml is sufficient sy present. I was planning to add this to the Brillo manifest once we have a board that uses it. Or should we go ahead and do it now for the benefit of the code release processes?

Patch-set: 1
Reviewer: Gerrit User 1080683 <1080683@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0