Update patch set 1
Patch Set 1:

> I was thinking of a process like this:
 > 
 > 1) Create separate manifest with just source projects for lldb
 > prebuilts
 > 2) Create builder for new manifest to build all of the prebuilts
 > 3) Someone manually downloads the compiled prebuilts, unzips them
 > into their respective prebuilts project, and uploads them
 > 4) lldb picks up the new prebuilts in the next build for lldb

This is the process I head in mind, except that I wanted to avoid the creation of a new manifest. The only reason for that was that I believed that it would require the creation of a new branch, which is quite heavy-weight.

If it is possible to create a new "prebuilts.xml" manifest in this branch, and make the builders use that, then that would be awesome.

(Another alternative could be to just add groups="prebuilt-sources" to the relevant entries, but again I am not sure if that would work. I believe you know better how the process works, so I'll go with what you think is best)

Patch-set: 1
Reviewer: Gerrit User 1058713 <1058713@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0