Update patch set 1
Patch Set 1: Code-Review+2

> This is the process I head in mind, except that I wanted to avoid the creation of a new manifest. The only reason for that was that I believed that it would require the creation of a new branch, which is quite heavy-weight.

There's no need to create a new branch for a new manifest. But you're right, I think we can avoid creating a new manifest as well. We still need a separate builder, I'm fairly certain we can have a builder for the same manifest but with a different entry script.

Patch-set: 1