Update patch set 1
Patch Set 1:

> There's no need to create a new branch for a new manifest. But
 > you're right, I think we can avoid creating a new manifest as well.
 > We still need a separate builder, I'm fairly certain we can have a
 > builder for the same manifest but with a different entry script.

Just realized there's a problem with this. There's no way to have the prebuilts builder *only* trigger on changes to the prebuilts sources. I.e., every time you push to external/lldb, it'll trigger a prebuilts build. This will be a massive waste of resources. We do need a separate prebuilts manifest that contains only the prebuilts sources, and not even the prebuilts binaries, since we're not checking those in directly.

The *branches* for the projects on that new manifest can all refer to lldb-master-dev, so that's not a problem.

Though, I think it would be neater if the branches and manifest names were the same, but since you already requested branches, requesting more just because I suck at planning ahead might be rude.

Patch-set: 1