Update patch set 2
Patch Set 2:

> > (1 comment)
 > 
 > What I'm planning to do is a git checkout SHA in the vulkan/glslang
 > directory. It would then no longer match the shaderc/glslang
 > version.

 > > (1 comment)
 > 
 > What I'm planning to do is a git checkout SHA in the vulkan/glslang
 > directory. It would then no longer match the shaderc/glslang
 > version.

You can't do that as part of the build. You need to set the revision="ref" attribute here if that's what you need.

Even so, I'm fairly certain the policy is that a project should only be mapped into the manifest once. +byi to confirm.

Patch-set: 2
Label: Verified=0