Update patch set 1
Patch Set 1:

> +srhines,pirama as an FYI. Our builds don't run the libc++ tests,
 > so this won't affect you. Some day treehugger will be affected by
 > this though, so maybe we should get a prebuilts/lit? Alternatively
 > I can bug ddunbar and see if getting pip's copy of LIT updated is
 > possible.

I guess the real issue here is where to constrain lit.py. If external/libcxx is allowed to move faster than external/llvm, I am not sure there is anything we can do to help (and perhaps there needs to be a distinct lit.py checked in for use just with external/libcxx).

Patch-set: 1