Update patch set 1
Patch Set 1:

This is the same clang gpu tools use for compiling their code, and I've been told it works ok on the buildbots. It has one major problem - the same clang does not generally work on the workstations, which have a new SDK, which it cannot understand. However, on the workstations, the system clang works fine.

It's not a very ideal situation, as it means we need to use different compilers locally and on the buildbots, but I guess it's not that different to the previous state, where we were using completely unrelated xcode versions.

Patch-set: 1