Update patch set 1
Patch Set 1:

> > This is the same clang gpu tools use for compiling their code,
 > and
 > > I've been told it works ok on the buildbots. It has one major
 > > problem - the same clang does not generally work on the
 > > workstations, which have a new SDK, which it cannot understand.
 > > However, on the workstations, the system clang works fine.
 > >
 > > It's not a very ideal situation, as it means we need to use
 > > different compilers locally and on the buildbots, but I guess
 > it's
 > > not that different to the previous state, where we were using
 > > completely unrelated xcode versions.
 > 
 > Do you mean that we may face some problems when using this clang to
 > build LLDB on our workstations - e.g., when running build-*.sh from
 > the shell?

That's true. I suppose I should add some ability to override the clang used via an environment variable or something...


 > This clang prebuilt version is 3.5 - can we make a new prebuilt
 > with clang 3.6 to use the same clang version both on linux and
 > darwin?

There is already a clang 3.6 prebuilt, but I don't know what's the state of it (i.e., locally it works as well as the 3.5 clang (=not at all), and i don't know how it would behave on the buildbot). I chose this one because I know it is already used by someone, so it should not be completely broken. I would not be opposed to trying out 3.6 though (maybe after trying this one first, as this is all still very experimental)...

Patch-set: 1
Label: Verified=0