Update patch set 1
Patch Set 1:

Most of the groups used in this manifest won't be used like this, but I don't see any other choice. (Most will be used to exclude certain areas, often for automated processes)

Can you file a bug on repo so that we can come up with a better solution?

Patch-set: 1