Update patch set 1
Patch Set 1:

I think if we reduce it down to just the groups listed below, everyone should be happy:

  adt-infra,cts,eclipse,motodev,pdk,tools,tradefed

Not all of those may be required, but it's a smaller list that includes the ones that are actually used independently (to my knowledge, when I wasn't sure, I included it).

The current list includes "notdefault", which would have broken the normal case of not specifying any groups.

Patch-set: 1
Label: Verified=0