Update patch set 3
Patch Set 3:

This is an additive operation now, so I'm not concerned if we need to add more internally. I took a quick scan of the internal manifest, and I'm not sure we'd need any changes there -- the group entry is only required here if it will be used to download only a subsection of the branch, so it's only a tiny portion of the groups that are actually defined.

Patch-set: 3