Disable merged commit pre-upload checks by default
When merge-commits are uploaded to Gerrit, the merged commits have
usually either already been reviewed elsewhere (in another branch), or
may contain automatically-generated commits that would probably not pass
the pre-upload checks (e.g. lacking a Bug: tag). This change makes it so
that the merge commits are not checked by the hooks, but the
merge-commit itself is.

Bug: 31926893
Test: repo upload
Change-Id: Ic25127fdbaea651cd4bb14e250c26ffe9aa064ad