Update patch set 1
Patch Set 1:

For on-device use, our current clients of SPIRV-Tools use it as a library, and we ship it in source form.  They are packaged together with the Shaderc sources.

I don't know how you intend to use it, but do you really need it as command line executables?

Patch-set: 1
Reviewer: Gerrit User 1095958 <1095958@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0