Update patch set 2
Patch Set 2:

Note: This CL caused some automerger issues, because master was at an empty commit internally.

We got errors that looked like "(fatal: refusing to merge unrelated histories )" because the internal master empty history was unrelated to the history on AOSP.

I went in and manually pushed to refs/heads/master (and the downstreams, but not the release branches), but this is something that maybe validation should catch?

Patch-set: 2