Use head version of qt on emu-master-dev branch.
There is a bug in qt which make modifier keys doesn't work well on os/x.
It got fixed in emu-master-dev with CL http://r.android.com/461216.
Since we don't have a emu-2.4-arc branch for qt, just change our
manifest to point to the version of qt with that fix on emu-master-dev
branch.

Bug: 62824830
Change-Id: Idc2eace84c7a01f43963e43a97a2389303d1a282
Signed-off-by: Tao Wu <lepton@google.com>