Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > > Patch Set 1:
>  > >
>  > > > Nested projects in manifests are not supported
>  > >
>  > > It works right now - toolchain/clang is nested as
>  > toolchain/llvm/tools/clang in this manifest.  Do you mean that the
>  > support might go way?
>  > 
>  > We've been removing all cases of this, since it doesn't work in the
>  > superproject world (the replacement for repo). It's also quite
>  > flaky right now in buildbot (though you're probably not seeing it
>  > on a low-activity branch).
> 
> The upstream LLVM build relies on this nesting, unfortunately.  We can workaround this particular nesting if necessary, but the rest of the projects (clang, compiler-rt, libcxx, etc.) need to follow their nesting.
> 
> This same setup is in go/ab/aosp-master-plus-llvm, which builds at the same rate as other AOSP builds*, but I don't recall seeing any flakiness there.
> 
> * for now - it'll be changed to trigger once every few hours in the future

Can symlinks work instead? That's what we've used in a handful of cases. In other cases we wrap the build to copy/etc the code into a temporary directory before building, but that's not a very good local experience.

It's possible that we'd be able to do nested superproject subscriptions in some cases (with nested git submodules). We're not claiming support for that in the general case, since it does introduce some limitations, and our infrastructure won't be expecting that. https://gerrit-review.googlesource.com/Documentation/user-submodules.html

Patch-set: 1
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>