Update patch set 1
Patch Set 1: Presubmit-Ready+1

We don't have a way to merge the manifest parts of this change, but it all needs to be "atomic". We know that it will likely have a build breakage until all of these CLs are in, but that is ok with us. Build folks, can you please merge this when you are ok with the manifest changes? Thanks.

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1