[WebView Support Library] format XML to reduce merge conflicts
Other CLs in this series: http://ag/4627933.

This adds three lines of non-changing comments before and after the
webview_support_interfaces XML line. The goal is to reduce merge
conflicts when cherry picking CLs from androidx-master-dev to
ub-supportlib-master.

No change to the pinned revision for the support interfaces, so this
should have no change to behavior.

Bug: 111757456
Test: N/A