Update patch set 1
Patch Set 1:

LGTM but I was going to hold off on doing this for just a bit until I can try adding a dependency on the project to ensure it's being built & included properly.

Specifically I'm going to have the systrace CTS test use this instead of it's current regex parser as that would be a solid use case to justify inclusion into the default.xml.

Patch-set: 1
Reviewer: Gerrit User 1010103 <1010103@85c56323-6fa9-3386-8a01-6480fb634889>