[DO NOT COMMIT] update manifest to use llvm monorepo
This is just an example of what the new manifest will look like.  We
will not fetch upstream directly from github, but are waiting for the
go/android-3p party to resolve.

Mostly, it's to show that there will be a new project, and the
fragmented sources will be dropped.

go/ag/7286501

Bug: 131714073
Change-Id: Ifce3b8c632e7b0e167bd06a4a7fc5f376ec53a91