Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > It seems like something is digging down into the directory structure and capturing unintended .mk files. I'm not sure how we want to proceed here. Dan/Colin, is there some preferred way to silence this for external projects?
> > 
> > toolchain/llvm-project/pstl/build/jni/Android.mk:9: error: proj_root: export is obsolete. It is a global setting. See https://android.googlesource.com/platform/build/+/master/Changes.md#export_keyword.
> > 00:15:49 ckati failed with: exit status 1
> 
> Add an empty Android.mk at the top level of the project (probably with a comment saying that it's intentionally empty to prevent loading deeper makefiles)

Nick, I think that is going to be the easiest answer for now. We should work on getting this done in upstream as well, so that we don't have to carry around a local hack (but the master-legacy branch has other hacks already, so ...).

Patch-set: 2
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>