Update patch set 5
Patch Set 5:

> Patch Set 5:
> (Seems like you have a build target for python before. Why is it not there any more?

I don't remember all the details. I think that we had a separate manifest for building prebuilts at one point (lldb-prebuilts), but it was in the internal repository, which I don't have access to at the moment. What I do remember is that building python on windows was pretty painful.

Patch-set: 5