Update patch set 1
Patch Set 1:

This will create a different problem, everyone who touches an Android.bp file will have to bpfmt it.  We would like to turn this on globally, but we need to bpfmt everything first, and we don't have the tooling to do that.

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>