Make it possible for everyone to create a change to manifest.
It seems that the restriction was initially added in 2012
(https://android.googlesource.com/platform/manifest/+/71e33f2bd92d62f62dbce9b98f15252908fb4544%5E%21/#F0 )
and we expanded access multiple time over past years.

Because we have control on +2 (which we also intend to migrate to
OWNERS model soon), it seems to be too restrictive and should be
removed.

Bug: 170431829
Change-Id: Id27cebc7a73f45b2892363bd9bec62cf147ce5a5