Update patch set 2
Patch Set 2:

Change bypassed Presubmit.

Bypass Reason: other 
Additional Feedback: Infra errors in presubmit and none of the errors point to this change.
In practice this doesn't even change anything because we're changing the manifest from a pinned SHA to ToT (which currently points to that same SHA).

Patch-set: 2
Tag: autogenerated:TreeHugger:Bypassed-Presubmit