Add submit requirements for project platform/manifest
We create a non-applicable submit requirement for each label function
that is empty, i.e. which does not contain any values. These labels were
by default hidden in the UI, hence creating a non-applicable SR to
prevent a parent from creating a SR with the same name making it appear
on the UI.

Change-Id: Ie1e38005867385c9e280f879f185f4aae7dc6bc7