Revert "Add external/rust/crates/hex to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (go/aca-get/cbce981d-842f-4a50-876b-ef3c43b4ac0a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic0a33d5de96a8789eecd6987ba613e1eefc0fab3