Revert "Add JDK 17 to androidx-main"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b01d943c-ce3c-4ded-ba2c-2a37ab05f826). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ice2239aaef05fb16d64a63fd1eea398367fe95b4