Revert "Add external/rust/crates/openssl-macros to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/acf332c7-1df8-41ec-b79c-e293f1df96c5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id9d4d76a3c3e8b29067ec7d9ea179d8cea43fc1a