Revert "Add NXP KeyMint and Weaver HALs to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/768c9160-e56d-40d4-999e-a6f874461728). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib20923b0530e386539bd7054c590883091ef5e6c