Revert "Add NXP KeyMint and Weaver HALs to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d550ffc3-a304-4724-a3f2-9ec06812bb72). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iee5fc5d3928545c119837721daec626bde490f7d