Revert "Add NXP KeyMint and Weaver HALs to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/46075022-b5ba-4d70-ae2c-45e6f660c45b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iddc8277bcaaa934a0ef2b4c607ec6ab44bc73763