Revert "Add NXP KeyMint and Weaver HALs to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a66c2df6-8190-4c93-949b-5cacbe64f2dc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iccae7c50c4a7865959975a4da876fa9da0a5a921