Revert "Add fastrand crate"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8865a8dc-9e89-444a-b268-a225b2f2c1e0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1518bef75e976d6ce2fb6ad3822afeacee2b9a58