Revert "Add NXP KeyMint and Weaver HALs to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4ceea7ec-486a-4e2d-adf6-5245b4164f4b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If0566883f5fdef26e1dc0857dad13707b4d49a19