Revert "Remove project"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/55050daf-e4d7-4d2e-9ddc-3ea2d1d111ca). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6d439eb6355d75e9dfb420d010397e94da8bac56