Revert "Add external/roboto-flex-fonts to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b3dce194-ed63-4a83-a216-4386370acb54). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5fd83f8c6c678b5badce7e885e6ee58d476d62f7