Revert "Add external/roboto-flex-fonts to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d086e4aa-006e-4442-a28e-2fceef804ae6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7b1d5f3e604469b4690ada08bef3365ca750e3df