Revert "Add external/roboto-flex-fonts to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/757eb4b2-ad5f-4a66-8841-22790cc8a0e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4987c72aab474cb0e98feff1bf17698917d180dd