Revert "Add external/roboto-flex-fonts to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/97539d99-ef7f-4a47-8d47-43dd0aac8b94). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id667b8e81fa11c8d0039eae47cb7188afaebbd1f