Revert "Add external/roboto-flex-fonts to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8265b75b-10a7-4d75-b704-412b3240655d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic4633c51e61917d53cec3897f5aa8581878bd562