Revert "Add external/roboto-flex-fonts to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/253fb7d2-61cf-450c-bc8c-c04ffb6049b2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I59e71c0ca7636d4d1c3047060d8d30959973cfba