Revert "Add external/roboto-flex-fonts to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1a04491b-7620-4890-a079-14b3fb4e94d9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Idab2a9660466375adce284a640d03fb5d82eb5ea