Revert "Add prettyplease for PDL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5902032e-23e7-42b6-b3c2-6cc9b7c537df). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I05a5bdf2e3323a673bc2e8ef4003146d80049db6