Revert "Add prettyplease for PDL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/18719de8-330b-49c3-a3be-fa28a269d7e4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0a5728154603fcf13ea9a33360270ae4b3b092a4