Revert "Add prettyplease for PDL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/887bb3c3-1b68-4747-9c27-ebcabc6d0ccc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib3bf211672e569e69cc2c8c16a4b74a5d619e7ff