Revert "Add prettyplease for PDL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/866286b5-38d8-45ce-b1de-99984709c4b9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie8b733d8f742aa91a70015d545da782bc4aa69f5