Revert "Add prettyplease for PDL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ba955242-46c8-4976-893e-d2e5b44513e8). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia3e09f8971435233682e09097af18933c3e19db4