Revert "Add prettyplease for PDL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/e8ca4882-2bdc-45fc-925f-76df8dfebf3e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6e1059f44fa3840e4883f25fbe45655a1aeeb800