Revert "Add prettyplease for PDL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5bf59cca-3416-4d77-9b23-2336e347cc46). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib792a418ed0264b9947d6084ce590a4276fb0cef