Revert "Add prettyplease for PDL"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2439bd55-e8a3-4a16-91e4-08e54f0d6e46). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia51b328fac4e4d17b18710121366bd757c48e581