Revert "Add external/jetpack-camera-app to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a6d82010-2ae6-400f-b636-eb28a13c0702). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I355f1fb40ac4b2d8d4397eec1ab40f068d68b077