Revert "Add external/jetpack-camera-app to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4a96e2bb-61b5-45e1-b1c0-71cf5ac4fdd4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4e75b30b8535a3aca064036eabf45de290b73f95