Revert "Add external/jetpack-camera-app to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/be39a8f1-fbbc-429e-8db9-4931dca4b3ef). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I17ffd49fb521e86160e7606abf14dc9c8d25962b