Revert "Add external/jetpack-camera-app to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/650d6e66-1a96-4670-a28d-260d6bd02d16). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia4af7ddf4b471e7725d1fd6f4e777bd4d716afc3