Revert "Add kotlinx.serialization to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3d8b6124-fd64-483d-ac5a-5a7d89d29856). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2ffc2cec33451d29abeb45eb4000fd14a64a1452