Revert "Add kotlinx.serialization to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f7ed59e0-878f-4b42-bec6-0c5c9e1d2c03). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I89f982b37ceb28b82b1782338c8ed399e0323376