Revert "Add kotlinx.serialization to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/72980a83-3a79-4913-b6b7-adc0c44f094e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I585cfb0188fa4c119cf510c5714d39036f02a33a