Revert "Add kotlinx.serialization to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/75b2e4da-983d-4210-af4a-88f6dc325b8d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic852a30237e88b2c21588ce3e56954aaf8352472