Revert "Add kotlinx.serialization to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/fe45a5df-e22e-4fa9-b5b2-70b5c239ef23). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1f34c826a1fc40f98d41fd1e4baac203fdd30167