Revert "Add kotlinx.serialization to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c5508569-78a7-4b2d-a758-9fcd94a7995e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id22120ba8802e589a0ce73ec6065de9a6db2363b