Revert "Add kotlinx.serialization to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/5661264e-436a-4704-ab0e-887c66106732). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00f80b5bcf657547e2e89d096b8eae63c1598cde