Revert "Add kotlinx.serialization to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a8365041-a33a-4106-8e0f-8ddcb9f3240e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I966cd4702ad52f52f551e66dd0dae3ce67328da8