Revert "Add kotlinx.serialization to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/627c43a6-46b3-40c7-b595-e8d97f7bda0c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4dcc651dc53ea142c19dda911805a3c1c9214e01