Revert "Add kotlinx.serialization to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/68c8aeaf-6d25-4455-a32a-07d1e5ec11ab). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0c18c8aea91ab9c639dbb107b4755424f727edc4