Revert "Remove jdk9 and jdk11 from manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d899cbf5-f9a7-4ff3-9166-24e77e34c197). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2544742f68eacf77427e0e2cc4a4b702aa0688e2