Revert "Add external/accessibility-test-framework to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/aa755186-a198-42a7-a23b-a98ed3b652d6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 343224728
Change-Id: I17b0af9a15ac52c5732fb159cc4da02668a768de
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com