Revert "Add external/OpenCL-CLHPP to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed528188-d796-4640-878d-9a90230dbe96). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 344765915
Bug: 344769767
Change-Id: Iaa1892982cbcc84b74ff090db3a79af56dcb1b88
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com