Revert "Add external/OpenCL-CLHPP to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7bd5a1fd-c1bd-4f86-9199-0eed3b0c84e1). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 344769554
Change-Id: I877337fb6c9828b6ef29cce4897ffd1266101cd2
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com