Revert "Add external/OpenCL-CLHPP to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/223c61c2-1c69-40bb-81cc-2e548660d0f5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 344769981
Change-Id: Ie6c7e73cee0a8b89701676f2188a8b655c396227
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com