Revert "Add external/OpenCL-CLHPP to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/432f969c-c955-41b4-be37-6d0a9df6805e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 344769038
Change-Id: Ida61dc0b3e1d551f53c5c952bb986ef470500622
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com