Revert "Add external/OpenCL-CLHPP to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/c9743672-90e1-4913-8599-9b2c5306800c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 344769668
Change-Id: If2ea31a699a8f1a88ac7eca27bdd64b39893e36a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com