Revert "Add tower crates to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83509966-dd29-423a-8a76-06fc8d067797). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 345858985
Change-Id: I3bc9e53c83cc24fb4aa334dfa4abebacb4430110
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com