Revert "Add tower crates to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/92106f3d-0fc0-453b-b7fe-263570bf6de5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 345859351
Change-Id: Ibfcf3cf92e9a53b7276e8a1d1132036291167328
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com