Revert "Add tower crates to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/83d73ddd-3f0b-406e-bd85-26a050d9e6a0). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 345859335
Change-Id: Ibac80f5396c5e1538be0510f3f9467b4b719a598
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com