Revert "Add "external/pciutils" to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8708e6ba-672d-439d-bff1-a491f0d335f2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 360792897
Change-Id: I194ea9af048e39bf358c3b1a40278e3f6fc4358f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com