Revert "Add "external/vboot_reference" to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8708e6ba-672d-439d-bff1-a491f0d335f2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 360792897
Change-Id: I96d85356dd5bb19cf642a6be2e757511c488e66c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com