Revert "Add "external/vboot_reference" to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4554e06-21d1-4ade-b8cc-5b6b14719dcf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 360797158
Change-Id: Ia1b835f2fbca740756e5c880836b49b744f7a45d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com