Revert "Add "external/flashrom" to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a4554e06-21d1-4ade-b8cc-5b6b14719dcf). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 360797158
Change-Id: Ieebf17ab21db405836a2910380ad29a5362b355d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com