Revert "Remove 112 Rust crates that have been migrated to the mo..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1d3f59b6-9358-4c15-a5ed-a70f19a8a7c5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362952251
Change-Id: I5383ec1d456554c7dfb7c73d21fd38dcd94cbb02
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com