Revert "Remove 112 Rust crates that have been migrated to the mo..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/bbf786df-f98c-4fe6-b831-8587424c4311). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 362952725
Change-Id: Id4478ccb868ca9923e5d11e60c8862d3a28788fc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com