Revert "Remove 112 Rust crates that have been migrated to the mo..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9d0a8158-5324-4ece-bd1f-a483987ab729). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icbc1f06a9927967b93fa6610f4f1489f62367795
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com