Revert "Add external/gmmlib to manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/74f31f39-9dec-41ba-b15e-f8808398caf2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 363000527
Change-Id: I0194dc9b80758d4b3ef9b9c36d459e943ca89299
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com