Revert "Add external/gturri* to asop manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/631e05b5-0bc2-4dd1-ac1c-0b858a96b5b9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 364994600
Change-Id: If37a4639060065a36433e6581f2a205481358a84
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com