Revert "Add external/gturri* to asop manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/781bafc6-2ec8-41ad-b972-55fb5a576a29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 364995260
Change-Id: I66a94b86f4b41941d6058b73266e7da4b9ad3fce
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com