Revert "Remove 112 Rust crates that have been migrated to the mo..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/972694e0-add3-487a-ab55-268c52048717). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1dcd311e2d3334cfc2da9f4105d1e6b495f44519
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com