Revert "Remove project tools/test/graphicsbenchmark"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/739880c7-a9d4-45c7-8568-dbf6bf56ce58). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 365632894
Change-Id: Ie583cf1859918849fd360af589e947ef745aca10
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com