Revert "Remove project tools/test/graphicsbenchmark"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1fb469af-a92c-4b48-8d81-3c42043dbc4a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 365676262
Change-Id: I6403255ec1a454ab81a2e1bb5ffadecf221f656d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com