Revert "Remove project tools/test/graphicsbenchmark"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4210ea06-26b2-4bad-a131-43edf06e44c2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 365694000
Change-Id: I2bd7bd45495a4d9871f55e9fd27f7c9a5ce15aac
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com