Revert "Add audio related projects"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ed31986d-b011-437f-8bec-dea06ecae2d3). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia226daf3ce20c7dd736ab4defa2f19a2d204d63d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com