Revert "Remove symlink to root Android.bp file"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cac634f0-bb1e-4991-a34f-340c026ae1ea). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I25df513471ff96757b94e7fe29b77a2d664df208
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com