Revert "Add json-schema-validator to ASOP manifest"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/1278434a-f2e4-4d1a-a5e6-b458b7d1dba5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 368059587
Change-Id: I23a65f08c65c8a565625dd81644d72e57c97f7eb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com