Revert "Remove unused device/generic/opengl-transport"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/696f209a-2bab-4001-a3e5-1877a89cd384). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I19c1da107735b1b2b41e7c6ddf399e9c06b65895
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com