Revert "Remove 20 Rust crates."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/50d64b4d-a669-4b50-aee6-aabfc5a4668a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 371066621
Change-Id: I718bb0e8b0b75bc32c951ab92c934cbda0090493
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com