Revert "Add external/rust/crates/v4l2r for crosvm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cdf6c015-db31-4648-a0c8-f114ad748b23). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 372352978
Change-Id: I10d760345f2f04d277190fd975e6fd64eecdedc5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com