Revert "Add external/rust/crates/v4l2r for crosvm"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0a3cc092-86d0-4f8e-977f-864f2ecf8bc5). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 372366193
Change-Id: I0349b9d10722c6a10c8d31437bb2e1f6c6585ee6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com