Revert "Add packages/apps/DeviceDiagnostics"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b313b643-b9c9-4774-9201-a8f6b070b715). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 372774461
Change-Id: I4e82ea6ef1627e14e51b2fd5ee1f1679817c254c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com