Revert "Add packages/apps/DeviceDiagnostics"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2138d3e5-a865-4e56-837f-cd069b03f38f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 372774736
Change-Id: I64f0cf69b1849501f77d624d14c7399f95cb1d4f
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com