Revert "Remove 20 Rust crates."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b6536193-16b8-4c38-b9f6-ffc1453ee994). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I4dc3acac4e05fea42c25002762acbc6375c1182b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com