Revert "Add pdk group to Trusty dependencies"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7793c9ee-f9d8-4f3d-b724-b339571853a4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 377832748
Change-Id: I6513f3dd121a562f9c1caf146dd793ff41a014af
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com